+++ composeApp/src/commonMain/kotlin/com/healthanalytics/android/App.kt
@@ -3,7 +3,7 @@ package com.healthanalytics.android
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.*
 import com.healthanalytics.android.di.initKoin
-import com.healthanalytics.android.presentation.dashboard.MarketPlaceScreen
+import com.healthanalytics.android.presentation.screens.marketplace.MarketPlaceScreen
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 // Initialize Koin when the app starts
