+++ composeApp/src/commonMain/kotlin/com/healthanalytics/android/presentation/screens/BiomarkersScreen.kt
@@ -13,6 +13,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.healthanalytics.android.data.models.Biomarker
 import com.healthanalytics.android.data.repositories.HealthRepository
+import com.healthanalytics.android.presentation.screens.dashboard.StatusChip
 import kotlinx.coroutines.launch
 
 @Composable
