+++ composeApp/src/commonMain/kotlin/com/healthanalytics/android/presentation/screens/LoginScreen.kt
@@ -15,7 +15,6 @@ import com.healthanalytics.android.data.repositories.HealthRepository
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
-
 @OptIn(ExperimentalMaterial3Api::class)
 @Preview()
 @Composable
@@ -26,8 +25,7 @@ fun LoginScreenPreview() {
             onLoginSuccess = { token ->
                 // Just log or handle the token for preview
                 println("Login succeeded with token: $token")
-            }
-        )
+            })
     }
 }
 
@@ -35,21 +33,18 @@ fun LoginScreenPreview() {
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun LoginScreen(
-    onLoginSuccess: (String) -> Unit,
-    repository: HealthRepository = HealthRepository()
+    onLoginSuccess: (String) -> Unit, repository: HealthRepository = HealthRepository()
 ) {
     var phoneNumber by remember { mutableStateOf("") }
     var otp by remember { mutableStateOf("") }
     var showOtpField by remember { mutableStateOf(false) }
     var isLoading by remember { mutableStateOf(false) }
     var errorMessage by remember { mutableStateOf("") }
-    
+
     val scope = rememberCoroutineScope()
-    
+
     Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(24.dp),
+        modifier = Modifier.fillMaxSize().padding(24.dp),
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.Center
     ) {
@@ -60,7 +55,7 @@ fun LoginScreen(
             color = MaterialTheme.colorScheme.primary,
             modifier = Modifier.padding(bottom = 8.dp)
         )
-        
+
         Text(
             text = "Enter your phone number to get started",
             fontSize = 16.sp,
@@ -68,7 +63,7 @@ fun LoginScreen(
             textAlign = TextAlign.Center,
             modifier = Modifier.padding(bottom = 32.dp)
         )
-        
+
         OutlinedTextField(
             value = phoneNumber,
             onValueChange = { phoneNumber = it },
@@ -77,10 +72,10 @@ fun LoginScreen(
             modifier = Modifier.fillMaxWidth(),
             enabled = !showOtpField
         )
-        
+
         if (showOtpField) {
             Spacer(modifier = Modifier.height(16.dp))
-            
+
             OutlinedTextField(
                 value = otp,
                 onValueChange = { otp = it },
@@ -89,54 +84,46 @@ fun LoginScreen(
                 modifier = Modifier.fillMaxWidth()
             )
         }
-        
+
         if (errorMessage.isNotEmpty()) {
             Spacer(modifier = Modifier.height(8.dp))
             Text(
-                text = errorMessage,
-                color = MaterialTheme.colorScheme.error,
-                fontSize = 14.sp
+                text = errorMessage, color = MaterialTheme.colorScheme.error, fontSize = 14.sp
             )
         }
-        
+
         Spacer(modifier = Modifier.height(24.dp))
-        
+
         Button(
             onClick = {
                 scope.launch {
                     isLoading = true
                     errorMessage = ""
-                    
+
                     if (!showOtpField) {
                         // Send OTP
-                        repository.sendOtp(phoneNumber).fold(
-                            onSuccess = { response ->
-                                if (response.status == "success") {
-                                    showOtpField = true
-                                } else {
-                                    errorMessage = response.message
-                                }
-                            },
-                            onFailure = { error ->
-                                errorMessage = error.message ?: "Failed to send OTP"
+                        repository.sendOtp(phoneNumber).fold(onSuccess = { response ->
+                            if (response.status == "success") {
+                                showOtpField = true
+                            } else {
+                                errorMessage = response.message
                             }
-                        )
+                        }, onFailure = { error ->
+                            errorMessage = error.message ?: "Failed to send OTP"
+                        })
                     } else {
                         // Verify OTP
-                        repository.verifyOtp(phoneNumber, otp).fold(
-                            onSuccess = { response ->
-                                if (response.status == "success" && response.data?.access_token != null) {
-                                    onLoginSuccess(response.data.access_token)
-                                } else {
-                                    errorMessage = response.message
-                                }
-                            },
-                            onFailure = { error ->
-                                errorMessage = error.message ?: "Failed to verify OTP"
+                        repository.verifyOtp(phoneNumber, otp).fold(onSuccess = { response ->
+                            if (response.status == "success" && response.data?.access_token != null) {
+                                onLoginSuccess(response.data.access_token)
+                            } else {
+                                errorMessage = response.message
                             }
-                        )
+                        }, onFailure = { error ->
+                            errorMessage = error.message ?: "Failed to verify OTP"
+                        })
                     }
-                    
+
                     isLoading = false
                 }
             },
@@ -145,14 +132,13 @@ fun LoginScreen(
         ) {
             if (isLoading) {
                 CircularProgressIndicator(
-                    modifier = Modifier.size(20.dp),
-                    color = MaterialTheme.colorScheme.onPrimary
+                    modifier = Modifier.size(20.dp), color = MaterialTheme.colorScheme.onPrimary
                 )
             } else {
                 Text(if (showOtpField) "Verify OTP" else "Send OTP")
             }
         }
-        
+
         if (showOtpField) {
             Spacer(modifier = Modifier.height(16.dp))
             TextButton(
@@ -160,8 +146,7 @@ fun LoginScreen(
                     showOtpField = false
                     otp = ""
                     errorMessage = ""
-                }
-            ) {
+                }) {
                 Text("Change Phone Number")
             }
         }
