+++ composeApp/src/commonMain/kotlin/com/healthanalytics/android/presentation/screens/dashboard/DashoardViewModel.kt
@@ -1,0 +1,4 @@
+package com.healthanalytics.android.presentation.screens.dashboard
+
+class DashoardViewModel {
+}
