+++ composeApp/src/iosMain/kotlin/com/healthanalytics/android/MainViewController.kt
@@ -1,6 +1,6 @@
 package com.healthanalytics.android
 
 import androidx.compose.ui.window.ComposeUIViewController
-import com.healthanalytics.android.presentation.dashboard.MarketPlaceScreen
+import com.healthanalytics.android.presentation.screens.marketplace.MarketPlaceScreen
 
 fun MainViewController() = ComposeUIViewController { MarketPlaceScreen() }
